Outsources exercise types
