+++ frontend/src/lib/Excercises/exercise.ts
@@ -1,35 +1,4 @@
-import type { languages, exercises } from "../constants";
-
-interface ExerciseType {
-  exercise_id: number,
-  exercise_title: string,
-  exercise_description: string,
-  exercise_type_name: string,
-  exercise_type_value: exercises,
-  exercise_language_type: languages,
-  exercise_language_name: string,
-  exercise_content: object,
-  exercise_solution: object
-}
-
-export interface ProgrammingExerciseType extends ExerciseType {
-  exercise_content: {
-    code: string;
-    func: string;
-  };
-  exercise_solution: {
-    key: [params: Array<number>, result: Array<number>];
-  };
-}
-
-export interface ParsonsPuzzleExerciseType extends ExerciseType {
-  exercise_content: {
-    list: Array<string>;
-  };
-  exercise_solution: {
-    list: Array<string>;
-  };
-}
+import type { ProgrammingExerciseType, ParsonsPuzzleExerciseType } from "./types"
 
 export const getExercise = async (
   exerciseID: number
